+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - 2019/08/27
 
     - fixed `-jobname` detection regex.
-    - added function `handle_jobname`
+    - added function `handle_jobname`.
+    - added the `--jobname` command line option.
 
 - 2019/08/26